+++ Policy/src/main/java/com/example/demo/repo/PolicyRepo.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.example.demo.model.PolicyModel;
 
 @Repository
-public interface PolicyRepo extends JpaRepository<PolicyModel, Integer>{
-  public PolicyModel findById(int id);
+public interface PolicyRepo extends JpaRepository<PolicyModel, Long>{
+  public PolicyModel findById(long id);
 }
